basic: HEX$(): force unsigned value (untested on BE hardware)
This breaks in hosted build, and judging from the implementation of
the ctor in BString, it should. But on BE hardware, it doesn't.
Hopefully it still won't after this fix.

+++ ttbasic/basic.cpp
@@ -4164,7 +4164,7 @@ BString shex() {
   int value; // 値
   if (checkOpen() || getParam(value, I_CLOSE))
     return BString();
-  BString hex(value, 16);
+  BString hex((unsigned int)value, 16);
   hex.toUpperCase();
   return hex;
 }
